Add work_assignment field, move config to new module
