+++ scripts/3.1_assign_primary_feasible_zones.py
@@ -13,9 +13,9 @@ from acbm.assigning.utils import (
     zones_to_time_matrix,
 )
 from acbm.cli import acbm_cli
+from acbm.config import load_config
 from acbm.logger_config import assigning_primary_feasible_logger as logger
 from acbm.preprocessing import add_locations_to_activity_chains
-from acbm.utils import load_config
 
 
 @acbm_cli
