+++ scripts/3.3_assign_facility_all.py
@@ -6,8 +6,8 @@ import acbm
 from acbm.assigning.plots import plot_desire_lines, plot_scatter_actual_reported
 from acbm.assigning.select_facility import map_activity_locations, select_facility
 from acbm.cli import acbm_cli
+from acbm.config import load_config
 from acbm.logger_config import assigning_facility_locations_logger as logger
-from acbm.utils import load_config
 
 
 @acbm_cli
